* update to SVN r32631.
git-svn-id: https://mameplus.svn.sourceforge.net/svnroot/mameplus@5191 e0215e0a-1e64-f442-9f1d-d2f61a7f4648

+++ trunk/mamep/src/mame/video/tc0150rod.h
@@ -18,7 +18,7 @@ protected:
 private:
 	// internal state
 	dynamic_array<UINT16> m_ram;
-	UINT8 *        m_roadgfx;
+	UINT16 *m_roadgfx;
 };
 
 extern const device_type TC0150ROD;
